Move ban classes to the ban package
