+++ src/main/java/org/spongepowered/common/ban/SpongeBanBuilder.java
@@ -32,9 +32,9 @@ import net.kyori.adventure.text.Component;
 import net.minecraft.server.management.IPBanEntry;
 import net.minecraft.server.management.IPBanList;
 import net.minecraft.server.management.ProfileBanEntry;
-import org.spongepowered.api.util.ban.Ban;
-import org.spongepowered.api.util.ban.BanType;
-import org.spongepowered.api.util.ban.BanTypes;
+import org.spongepowered.api.service.ban.Ban;
+import org.spongepowered.api.service.ban.BanType;
+import org.spongepowered.api.service.ban.BanTypes;
 import org.spongepowered.common.SpongeCommon;
 import org.spongepowered.common.accessor.server.management.IPBanListAccessor;
 import org.spongepowered.common.adventure.SpongeAdventure;
